Refactoring markerColors as a service from app context
